[DAGCombiner] refactor select-of-FP-constants transform
This transform needs to be limited.

We are converting to a constant pool load very early, and we
are turning loads that are independent of the select condition
(and therefore speculatable) into a dependent non-speculatable
load.

We may also be transferring a condition code from an FP register
to integer to create that dependent load.